Fix apparently longstanding FORM OBJECT! bug
Forming of an object would print the object line by line, each with a
newline, and then remove the last character in the mold buffer to get
rid of the last newline.  There was no accounting for if nothing had
been added.

It's strange that this was in the 12-Dec-2012 open source release and
it hadn't been a (known) problem yet...seems like it would be common:

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/core/s-mold.c#L880