Fix release build APPLY regarding typeset symbols
In the elimination of UNWORD and use of typesets for frame keys,
the symbol position was deliberately different from that of the WORD!.
This is tracked by a debug check to make sure typeset access to the
symbol is not used on a word and vice-versa.

However, APPLY obeys the debug/release switch for LEGACY options
enablement.  It had a reference to VAL_WORD_SYM that did not get
checked by the assert, leading to a bad symbol being used for
refinements instead of the word symbol of the refinement itself.