Eval_Table skips unnecessary evals, ~6% (?) speedup
The evaluator loop is fairly optimized in the release build, but it
still costs more than not calling a function at all. This makes a new
note in the type table about whether that type has any behavior in
the evaluator at all... and that is used to generate a small TRUE/FALSE
table of whether each REB_XXX type needs the evaluator. This is
used in a new macro implementation for DO_NEXT_MAY_THROW in
order to bypass evaluation altogether.

The test isn't actually as simple as just checking the next value, due
to the behavior of infix. Even `foo [x] + [y]` might have some meaning
for an infix operator. Hence to take advantage of the skip there must be
*two* inert values in a row, like `foo [x] [y]` or `foo {a} 2`. (For these
purposes, REB_END counts as an inert value, so `foo [x]` is also able
to be optimized.)

Casual performance testing showed that this is beneficial enough to be
worth doing, as a 6% or so speedup in building hostilefork.com. Further
tests would have to be done to know how typical or atypical that is.