OPTIONAL => OPT, RELAX => TO-VALUE
In reviewing some of BrianH's lingering PRs on GitHub, he had used
the term TO-VALUE for the concept I had called RELAX.  It is probably
a better name, especially now that unset is not considered to be in the
ANY-VALUE! class.

This also accepts that OPT is shorter and learnable for the purposes of
turning NONEs into UNSET, and deletes the longer name.