Add SET? native, remove x: () disablement from legacy
As the meaning of the LEGACY switches is formalized, they are to
help with porting code from R3-Alpha to Ren-C...and are not there
to force the Mezzanine to be able to run under modes for reasons of
"hypothetical need".  This moves the cant-unset-set-words switch
from being one that the system is maintained to be able to run with
in effect to one that is documentary in purpose only, as code that
was dependent on unset assignments being an error is not common
enough to be limiting concern on usage of unset assignments.

Also adds a SET? native to complement the UNSET? test.