Fix LEGACY(OPTIONS_EXIT_FUNCTIONS_ONLY)
The definitional return change meant that RETURN and EXIT were
fundamentally different operations, and RETURNs were expected to
hold the special "trick" of the function frame they definitionally
returned to.  The behavior change where EXIT would lead to returning
from the top level scope was based on throwing a return instead of
an exit, and this meant in legacy mode it was throwing a corrupt
return.  This moves the behavior of the switch to the location of the
catch of the EXIT at the top level.