Zero-terminate BINARY! made from CHAR!
Doing TO-BINARY! TO-CHAR 65 would result in producing a non
zero terminated data series.  While technically speaking there may not
be much need to zero terminate a binary blob, the original comment on
the open-sourced code for Make_Binary() said:

    Make a binary string series. For byte, C, and UTF8 strings.
    Add 1 extra for terminator.

A more aggressive assert recently started checking this each time an
ANY-SERIES! was initialized from string, and is catching occasional
violations.  This is another case.

Broader review is probably needed to see if it's better or worse to have
the terminator for consistency with the other series types.