Add SOMETHING? helper (not unset, not none)
There are some cases where testing for either NONE or UNSET is
what you want.  This adds a basic base-defs definition to the bootstrap
and a higher-level definition to r2r3-future.r.

In particular, this can help deal with unused refinement arguments in
the changeover to when they are unset in Ren-C, vs. NONE! in Rebol2
and R3-Alpha