Permit URL! to be converted to TUPLE!
PARSE of a URL! in the network code was capturing a range out of
the content in order to locate the host name.  If it were numeric as in
`tcp://93.184.216.34` the host name would be converted TO TUPLE!
to signify it was an IP-address.

This was not working, because R3-Alpha would only convert STRING!
and the bug went unnoticed because the conversion was inside of
an ATTEMPT.  This changes to accept URL! as the "easy fix", but
adds notes on the matter for deeper consideration.