Rename Qualify_String, factor out scan constants
The routine Qualify_String had a cryptic name, and some instances
would comment what it did...warning readers that it could raise errors
and that the buffer it returned was temporary.  Those two facts were
much more important in the name than "Qualify" or even "String", so
this renames it to `Temp_Byte_Chars_May_Fail` and expands upon
the comments for the reason for its existence.

There were also several hardcoded constants used with this routine
to make buffers such that they would align with the maximum length
of a scanned input for a type.  This puts all those numbers together
as constants in sys-scan.h so that they may be considered (in
particular, how necessary it is to try and codify such limits for these
conversions, as it seems error-prone to keep in sync with the
scanner if new notations arise.)