Add ASSERT_SERIES to Debug_Series
No series that is not valid should be passed into Debug_Series, as it
uses the same printing routines as the mainline code.  Given that it is
likely of high interest in debug builds if one is outputting a series to
inspect it to know about any anomalies, this adds an ASSERT_SERIES
to the routine.

It also takes the const off the series argument.  Although REBVAL const
correctness has been generally reached, correctness on REBSER has
not become a scheduled realistic near-term goal...so having a single
instance that also needs to use a mutability cast isn't of much value.