Update http error message for "working error world"
The code comments and seeming design of errors looked to be designed
to enable custom fields.  However, this feature was not working.  If it were
working and the error were considered to be built implicitly inherited from
the error object template in %sysobj.r, then with that template containing
message it would override an outer variable named message, as in:

    >> foo: "Outside"
    >> base: make object! [foo: "Inside"]
    >> derived: make base [bar: foo]
    == make object! [
        foo: "Inside"
        bar: "Inside"
    ]

The long-term of the error design is not fully planned, but enabling it
to inherit caused that bug with the "message" template field.  This
renames it to "msg" as a workaround with a note on the problem.