Starting overhaul of errors incl. "self-awareness"
This change brings about new flexibilities in the error design as
well as some new features.  It is large in terms of number of lines of
code, yet many of those lines of code are asserts and checks to make
sure the assumptions are correct.  Because the tests provided the list
of what most of those assumptions were, it did not change them and can
pass core-tests.r with nearly no modification...in both legacy mode and in
normal mode.

Key features include eliminating the fixed idea of `arg1: arg2: arg3:` as
the parameterization, letting you define named properties of an error
object that are more meaningful.  It also lets you define a template that
will do substitutions when the error is printed, as system errors could:

    >> err: make error! [message: ["the" :animal "has claws"] animal: "cat"]
    == make error! [
        code: 1000
        type: 'user
        id: 'message
        message: ["the" :animal "has claws"]
        near: none
        where: none
        animal: "cat"
    ]

    >> err/animal
    == "cat"

    >> fail err
    ** user error: the "cat" has claws

It also has provisions for the internal code to be conscious of how many
arguments are in an error template and check this automatically in a
call.  This should replace the `Error_0, Error_1, Error_2`... macros in the
C code (in an upcoming commit) and be more robust.  The names of the
arguments in the templates could also be changed, though for the moment
they are still left as :arg1 :arg2 :arg3.

As an added benefit it has output showing the source file and line number
in the C code (in the debug build) for any error originating from source.  This
can be seen directly in the error, or if you didn't probe it as an object but
were just notified the error through a failure message, the info is added to
the `why?` command:

    >> trap [1 / 0]
    == make error! [
        code: 400
        type: 'Math
        id: 'zero-divide
        message: "attempt to divide by zero"
        near: [:value1 :value2]
        where: [action! / trap]
        __FILE__: %../src/core/t-integer.c
        __LINE__: 359
    ]

    >> 1 / 0
    ** Math error: attempt to divide by zero
    ** Where: action! /
    ** Near: :value1 :value2

    >> why?
    DEBUG BUILD INFO:
        __FILE__ = ../src/core/t-integer.c
        __LINE__ = 359