Module unset patches, some for Needs header
These fixes partially pertain to an easy-to-make error that should be
watched out for when using refinements via get-word intermixed with
"unset-revoked" refinements.  When considering either of:

    foo/bar arg :bar-arg
    foo/:bar arg :bar-arg

The answer is you should only do `foo/:bar` if you have some condition
where the argument is not there, e.g. if you are composing the argument
out.  So in the above, it should just be an ordinary `foo/bar`.  On the
other hand, this could be correct if bar is either the word bar or none:

    do compose [foo/:bar arg (:bar-arg)]

The reason is that the compose would have elided the bar argument if it
were unset.