Move IF/ONLY and relatives to r3-legacy
Though being a suggestion from myself, IF/ONLY and its friends
have not turned out to be all that popular.  In addition to being used
very rarely, they present a barrier to optimization in the evaluator,
which becomes more pressing as the unset changes open doors to
the use of IF for rapid chaining and "opting out".  Should IF, UNLESS,
and EITHER be constrained to not having refinements, and join ANY
and ALL...then these natives can be folded into the evaluator loop
sans call frames.

Syntactically, the construction `if a [[b]]` may be contracted down to
`if a '[b]` under future ideas of "lit bit".  In the meantime accepting the
idea that the branches don't have to be blocks is an overall "win", so
that change that came in the same commit will be retained as this one
is removed.

Support still exists in the `do <r3-legacy>` mode.