Reap benefits of self-aware errors
This commit is expansive and looks large, but simplifying.

Before the error system knew for each error constant how many template
arguments that error had, a variadic error creation function was not
feasible.  Hence macros like Error_0, Error_1, Error_2 had to be made.
Once the system could understand %errors.r well enough to check it, a
true variadic Error(...) function could be written.

Debate on the word to trigger errors originated for a while, apparently
starting with `cause` and then moved to `raise`, ultimately settling upon
simply `fail`.  The advantage of `fail` is that it suggests an error condition
by its very nature, while `cause error` and `raise error` suggest the need
for another component to make the suggestion.  For the same reasons
the native was called FAIL the internal function was renamed thusly.

The "cute" hack to simulate a keyword was removed, in favor of parens.
So what used to be:

     raise Error_2(RE_ARG2_ERROR, arg1, arg2);

Is now:

    fail (Error(RE_ARG2_ERROR, arg1, arg2));

The belief being that the "keywordishness" and stand-outness can be
suggested strongly enough by a lowercase short name and a space,
without the penalty of the problems that the "trick" to omit the parens
cost.  The readability improvements by no longer needing to mention
the arity of the error are thought to outweigh the loss.

This shift cleans up concerns about design first voiced by
@ShixinZeng, and undoes the previous trick in favor of a more pleasing
trick, which has provably tested well on early boot panics reporting
errors in an informative way.  So during an early boot:

     panic (Error(RE_ARG2_ERROR, arg1, arg2));

In that expression, the core error generator is the one who knows it's
too early to report an error...and actually exits out the process, without
ever making it to the panic call.  This consolidates the error reporting
mechanisms during early and late boot to a common pattern.

The undoing of the prior hack means that the error generation
functions are themselves reusable, and do not terminate by their nature.
They return frames and not values, which also streamlines the trap
mechanism in cases where values are not needed.