Fix mold of an Image with length of multiple of 10
If the length of the image is multiple of 10 (e.g. 5x1), mold of it
will be like:

make image! [5x1 #{
RRGGBBRRGGBBRRGGBBRR
GGBBRRGGBBRRGGBBRRGG
BBRRGGBBRRGGBBRRGGBB00
}]

while expecting:

make image! [5x1 #{
RRGGBBRRGGBBRRGGBBRR
GGBBRRGGBBRRGGBBRRGG
BBRRGGBBRRGGBBRRGGBB
}]

The problem is that an extra newline is counted in this case. Only the
newlines between the first and last pixel is needed, if the last pixel is
right at the end of the line, no need to count it here, as it will be
appended later.