Allow seeding of evaluator with value not in series
Historically, the DO evaluator would take a series and an index offset
into that series.  The first step it would take would be to pick the value
off at that index, and then proceed along from there.

This moves the first step of picking the value out so that the initial
value comes from the caller.  While most clients will treat this the same,
it makes it possible to simulate a head value for an evaluation that is
not resident in that series.  e.g. if one had the path `append/only` in
one location, and the series `[[a b c] [d e]]` in another, one would not
need to splice them together in order to `append/only [a b c] [d e]`
because that first value can be "pre-imaged" in the position at 0 cost.

The functionality is now used by Ren-C++, to avoid doing temporary
insertions or creation of temporary series when the sole desire was to
effectively get the evaluator to do an "apply-like" operation without any
series mutation cost.  It is likely to have applications within the Ren-C
code as well.