Move APPLY to user-mode, legacy
There have been considerable enhancements to the evaluator--with
features like unset arguments to refinements revoking the request for
that refinement--and get-words or parens able to be used to evaluate
a refinement within a path.  These enhancements provide a general
replacement for the rather brittle positional mechanisms of APPLY.

Advantages to removing the now-deprecated APPLY implementation
out of the C code to userspace are numerous.  It means that there
can be a focus on just one evaluator, that can be tuned and sped up,
without worries about the numerous bugs that result from having two
parallel C codebases that must be adjusted when the bit-level needs
of the evaluator change.

This user-mode apply does not support infix functions.  It could be
modified to do so, but as it is being deprecated in the first place the
code complexity is likely not warranted.  It will certainly be slower,
but the correct response to that is to use the new mechanisms.

https://trello.com/c/46zJ9lZj