Canonize words just once for refinement check
In order to do case-insensitive comparisons, the two symbols to
be compared must be first "canonized".  This is done by the macro
SAME_SYM(sym1, sym2) for both sym1 and sym2...then the canon
symbol numbers can be compared directly.

Previously when searching refinements, the use of SAME_SYM
meant that a word requesting a refinement in a path could potentially
be canonized several times during the same call.  This does that once
at the time of pushing that word to the stack.

A similar "one-time" canonization was considered for the paramlist of
the function as well...but rejected because it's preferable for word-valued
refinements to have the same casing they were written using.