Switch evaluator to using named DO_FLAGS
The evaluator had two flags as arguments, one for "lookahead" and another
for "next" (as opposed to the ordinary behavior of doing to the end).  These
were passed as booleans and occupied two flags.  By consolidating them as
bit flags they take up less space and can be extended if other alterations to
the behavior are requested.

The flag pairs are specified explicitly even though they are mutually exclusive,
for readability at the callsite and as a sanity check.