Semi-audit of path evaluation for THROWN, bugfixes
In the general audit of places that could potentially throw, the path evaluator
got kind of skipped.  It would appear that paths--not always allowing parens--
did not have a good hook point for bubbling up the thrown values.  There were
a few other problems in that area.

This reduces the interface to the essential information, and also pushes the
evaluation of refinements into Do_Path as opposed to being a burden on the
evaluator core.  This means there is never a way to get back "the unprocessed
part of the path" from Do_Path... it always "finishes".  It simplifies the interface
to not require an in/out parameter for the path, and puts pressure on conditions
that might have swept errors under the rug (or quietly discards throws).