[AGG] Fix multiple grad-pens in one draw block
Think of the following draw commands (psedu code):

grad-pen grad1
box ...
grad-pen grad2
grad-pen grad3
box ...

When agg_gradient_pen is called on grad2, box is not drewn yet, it was
only queued into m_attributes array, so m_color_profile (same as grad1)
can't be deleted. While it's processing grad3, as m_color_profile will
be pointing to grad2, which is not referenced by anyone, it has to be
deleted or it will have a memory leak.

If m_color_profile is being referenced anywhere, it's gotta be the last
attribute, so just checking the equality will tell if it needs to be
deleted before creating new storage.