Incorporate test source analysis, correct id-mismatch
This does a submodule commit of the changes which bring a Rebol-based
C parser to bear on a source analysis framework for the code.  It also
corrects two easy issues found by it where the name in the header comment
for functions did not match the name of the C function being documented.