Remove asserts from DROP_TRAP
Over time, the usefulness of treating DROP_TRAP as a
"checkpoint" where the state is assumed to match the push it
corresponds to has fallen off.  The issue is that you might have
some outstanding work on the stack level you were working at
when you did PUSH_TRAP, that you would like to cleanup after
the DROP_TRAP but before that stack level is popped.

This removes the last few asserts and acknowledges that it's a
better idea to check these in things like the outer evaluator loop.
Those checks are already there and will catch the problem without
getting in the way of mixing error handling with unfinished work
within a state.