DS_ARGC => D_ARGC, DSF_NUM_ARGS => DSF_ARGC
The DS_* naming convention was old and seemingly there was one
straggler from it.  This fixes it to the not-necessarily-great but long
standing convention of D_ prefixes in natives and actions to pick up
arguments out of the current call frame.

It also renames the DSF_ version (that is explicitly parameterized
with a call frame) to match.