Separate out Is_Value_Managed from the assert form
Previously there was an assertion called ASSERT_VALUE_MANAGED
which would check values that had series in them that might be managed
and make sure they were if they could be.  This splits that debug-only
function out into an ordinary function for testing the condition, which is
then used by the assert.