Call frame uses symbol ID, not WORD! REBVAL
In the past, there was a binding stored on the REBVAL word in
call frames--which is no longer necessary.  Some reduction of use
of words is already underway in the Do_State to just a label_sym,
and in the interests of a coming unification of the Do_State with the
fixed-size portion of a call frame...this reduces it to store just the
symbol ID as well.  So it's DSF_LABEL_SYM as opposed to the full
DSF_LABEL word.