add length and null check to ASSERT_FRAME
The ASSERT_FRAME function was walking through the key and value
lists and making sure they had a REB_END at the same position, but
never checked the tail count for matching lengths.  This adds the assert
along with a message on failure, as well as a message instead of a
crash on null keylists.