Unify fixed size part of call frame w/Do_State (BIG!)
This is a sweeping, "simplifying", and broadly enabling change that
eliminates the overhead of proxying state between the internals of the
evaluator and a separate idea of a "call frame", by making the fixed
size portion of all call frames equivalent to the input/output/state vars
of the very Do process itself.

In addition to the changes, there are copious comments explaining
what the parts are for and why they are there.  The best place to look
is at the comments on `struct Reb_Call` and its fields, which are laid
out in great detail.

This is the beginnings of the payoff of untangling the logic of the Do
evaluator into a single function with no repeated portions.  With a clear
view of the footprint, the need for a separate call frame type disappears,
and only an ordinary array is needed.  Because REBVAL arrays are
allocated from a pooling structure with basically all the benefits of the
chunked allocator, this gets rid of the chunked allocator and just uses
a call to Make_Array.  This array can then be either kept manual and freed
(as with a NATIVE!) or managed and handed over to GC--with no need to
proxy values into a separate object for a CLOSURE!  The net is a slight
performance gain with much less code complexity.

Yet the study of the now-unified structures and their fields organized
them such that parts which are not needed during a function call could be
used by the natives themselves, or even by the non-native dispatch code.
The expansion gives them 4 pointers which can just pointed to any value,
and if it is managed then that pointer will protect it.  It also offers an added
free GC-protected REBVAL cell to be used.  This is done by piggy-backing
on the sunk cost of participation of a running call frame with the GC.

In the process of establishing assertions and invariants for that service, a
bug was discovered through stress testing that `eval` was allowing user
functions to be potentially garbage collected.  This was solved by having
a temporary value internal to the Do evaluator which will only become
GC guarded when an eval is encountered, and then unguarded after that.

All told, this is a huge and promising change--there will be much to show...