Generate native prototypes in tmp-funcs, reb=>c update
All the native functions in Rebol have real C functions behind them.  So
in theory you would be able to take the address of one, and compare to
the embedded pointer in a native value to identify the native:

    if (IS_NATIVE(v) && VAL_FUNC_CODE(v) == &N_parse) {
        /* v is an instance of PARSE */
    }

In practice the native prototypes were not part of the system-wide API,
and the definition of N_parse was only visible within u-parse.c (at the
point in the file after it was defined).

This moves the code for generating the natives list to add them to
tmp-funcs.h.  Subsequently it streamlines the previous technique for
identifying natives, which required capturing values into the root object
during boot time.  It also leverages the ability of natives to know their
own REBVAL while running via access to the call frame.

Adjusting the code required moving the function that turned Rebol word
spellings to the more limited C-compatible identifiers into a common
location.  This enhances that function as well, e.g. the symbol for a
hyphen was turned into `SYM__` previously but is now SYM_HYPHEN.
It also changes it so that where `and~` mapped to `and` before, it now
maps to `and_t`, and includes underscores for readability so that e.g.
`infix?` becomes `infix_q` instead of `infixq`.

So long as lines in switch statements were being touched for the renaming,
this formats them to make it easier to see which are fallthroughs and which
are not, and removes the need for horizontal tab points of alignment by
using vertical space as is "Ren-C standard"