Experiment to fix misconception on const ptr optimization
The following does not--despite what @HostileFork thought--create a
no cost pointer alias in optimized builds:

     REBVAL * const foo = D_ARG(1);

That pays for both a variable and an assignment which does not get
optimized out.  A real zero-cost way of doing it is defined in this commit,
and it looks better:

    PARAM foo = {1};
    REFINE bar = {2};

    if (IS_INTEGER(ARG(foo)) && REF(bar)) { ... }

Though it is a bit annoying to have to use the macros on each access,
there's not really any feasible way to avoid it and manage to directly use
the variables in the call frame.  On the plus side of such an annotation,
it can help serve as a reminder that values in these locations are protected
from GC.

To further accelerate the performance, this caches a pointer to the head
of the block data in the call frame.  It reuses the existing `arg` pointer
which was not in use during the function call anyway.

It's applied to just a few cases that were being studied, and will be used for
a while before doing a system-wide hunt to search and replace using the
technique.  That will give it a chance to sink in and see how it might be
improved further.