Implement "extra" allocation for deep copies
There are several variants of the series and array copying routines,
and one of the aspects is the ability to make a copy but ask for an "extra"
amount of capacity in the allocation.  This feature was not implemented
for REBVAL-bearing arrays that were being copied deeply.  Hence an
append to one of these copies would likely cause a series expansion,
even if the knowledge of the need to do the append after the copy
existed at the time of the copy.

This adds the "extra" version for that case, which is notably used by the
function generators when they add a `return:` onto the specs.