`native` and `action` natives defined in C comments
The bootstrap process has two "special" natives that have to be set
up before the other natives constructions could happen.  These were
in a file called %booters.r and did not fit the usual construction process.
Consequently, their comment specs looked like:

    //
    //  native: native none
    //

This pushes the process through so that not only can NATIVE and
ACTION have their specs defined in the C source, but also they can be
defined to use refinements.  To take advantage of this, a previous hack
of integer values in the spec was removed, and ACTION/TYPECHECK
was added to pass the integer in specifically as an argument.  Then
the actions for doing type tests were promoted to getting official action
numbers equivalent to the type.

Beyond the general source aesthetic benefits of this change, it also
moved the declarations of the type checks into Rebol instead of in C,
making it easier to give each type check a custom description string.
(as opposed to "checks for this type" which would not be very helpful
if the function is in an anonymous context.)