Update parameter unpacking experiment macros
This takes a step further with the parameter macros so that they
actually form the entire declaration and initialization.  Consequently
the debug build can actually do a full cache of the argument value
pointer and the state of the refinement upon entry, which is helpful
for inspection.

It also eases the restriction of using ARG access on refinements, as
they are args and their slots can be reused.  However, REF can still
only be run on REFINE-declared args.