Simplify/cheapen SER_POWER_OF_2 logic
The SER_POWER_OF_2 flag was added in order to catch the
case where a series whose elements did not divide evenly into a
power of 2 were unable to use their actual series length as a means
of knowing their actual allocation size--throwing off the measurement
of the total amount of system memory in use.  (Rebol's allocator does
not remember the size of allocations, except to check that the right size
is passed in to free in debug mode.)

This simplifies the answer to the question of the "actual allocation size"
by removing the need to come back with the answer of how big the
memory pool was (it's not necessary to know; the pool takes care of
that) and reduces it to just maintaining the flag on those series which
do not fit evenly into the width.  It also removes the alignment of the
allocation size, leaving that to the responsibility of the allocator alone.

(e.g. rather than `malloc(31)` it would ask for `malloc(32)`, but a good
allocator should not be second guessed on something like that, especially
given that it interferes with bounds checking.)