Make DO "tick counter" thread global, SPORADICALLY()
The evaluation counter in Do_Core has been very useful for dealing with
bugs that happen at the same tick in a reproducible session.  This moves
the tick counter into a thread global so that it can be used for other
purposes, such as alternating the technique used for a given calculation
between debug and release methods on every Nth tick.

(Though such a process introduces a bit of "randomness" it's better than
waiting until one does a release build to find out about a path that has
been neglected...)