Resurrect chunked stack as value-array service
The chunked call stack was introduced to solve a problem with
R3-Alpha.  It had used values pushed to the data stack in order to
hold call frames for most functions (except closures, which used an
object frame in order to be persistent outside of the boundaries of a
stack structure).

This was not suitable because the pointers in the data stack could
be relocated in the event of a stack expansion, and natives were
coded to trust the persistence of pointers to values in their call
frames and not re-fetch them on every use:

    REBVAL *thing_i_hold_onto = D_ARG(1);

The chunked call stack was conceived as a reasonably performant
and simple way of getting a stacklike behavior for call frame memory.
Because it wasn't a REBVAL-specific structure, it could be
specialized to the actual pointers and types that it was talking about,
and not have to poke things into secret random fields of REBVALs to
save on space.

With the unification of the DO state and the call stack, it was no longer
needed to have a chunked allocator for most of the call frame data.
The fields were the actual state variables of the DO invocation on the
C stack itself, with the frames linked there.  The first take on this idea
was to make the variable portion of the stack frame (the REBVALs of
the frame arguments) always use a series allocation, as those are
already pooled.

Testing showed that this cost significantly more than the chunked stack
had, and the formerly-faster-than-rebolsource Ren-C dropped to being
slower.  In light of this the chunked stack is back but in new clothes--
this time it only stores REBVALs and is a generalized system service.
Because it's disentangled from the call stack mechanics, it's possible
to use it for any other stacklike allocation of value arrays that might be
needed...and they also get garbage collector coverage.  Unlike with
the data stack, these values do not move--and hence can be used as
the `out` target of an evaluation, or other such purposes.