Relocate assert regarding THROWN in Do_Path
Attempts to nail down what exactly the interface contract is for
Do_Path led to an assert regarding the throw state of the output,
which was being checked before an assignment that indirectly did
the update of the `out` through a stored pointer.

Though Do_Path continues to be targeted for a general rewrite,
this does the simple thing and moves the assert to after the update
has been performed.