Simplify Error_Arg_Type, quote errors at END!
Due to the "quoted unset hack", it is possible to make functions like HELP
that will be able to run with reduced arity (e.g. no arguments) even though
they are able to take arguments.  In order to get the behavior, one must
have a terminal argument which is quoted *and* that argument must be
explicitly specified as allowed to take UNSET!.

Because of this feature, quoted arguments would go through traditional
type checking on unsets and then at the end provide a message saying
that they did not take UNSET!, if the end was reached.  While this does give
some "discoverability" to the hack, it's confusing because it is a different
message than you would get if you reached the end of a block evaluation
and the argument wasn't quoted.

This adjusts the situation so the same message is given, which buries the
existence of the arity hack a bit deeper for the sake of uniformity.  Hence
typing something like `do [comment]` will say that comment is missing its
value argument (instead of that unset! is not a valid type for comment's
value argument).