Make COMMENT the first "frameless" native
In the split of responsibility within R3-Alpha, the core evaluator loop
would build a "call frame" for a function with all of its argument values
ready.  This involved an allocation of space to store the argument values,
calling into the evaluator, and checking the types.  The native would only
be called if all the prerequisites were met.  Then if the native completed
successfully, the allocated space for the frame would be freed.

Ren-C corrected a flaw in R3-Alpha's strategy for allocating these frames.
The flaw is that they were placed inside of a Rebol series which could need
to be expanded, invalidating the value pointers inside of it.  Rather than
use an ordinary memory allocation, it kept matters efficient by allocating the
values from a "chunked" memory manager...which was relatively fast to
allocate and free from.

Yet still faster than either method is not to do any allocations or frees (even
cached/pooled ones).  The call frame method was redesigned in Ren-C so
that the "do state" (local variables to Do_Core) were packaged into a structure
which became *the call frame itself*, so that it would be possible for a native
to be styled in such a way that it could run with access as an extension of the
evaluator itself...prior to the point of any frame allocation, type checking, or
freeing.  Some natives are particularly suited to this "frameless" style and can
perform their work without ever allocating the variable-sized REBVAL array
that brings some cost to invocation.

As the simplest imaginable first case study, COMMENT was implemented as a
frameless native.  This means that it's able to take into consideration the block
position and index, evaluate and update that index, react to any type errors or
seeing the end-of-block, etc.

This raised a few issues.  One is that all frameless natives will need to have a
framed implementation as well, for debugging introspection.  If there is no frame
then there is no "standard API" for getting insight into what is going on when
browsing the call stack (it's a bit analogous to seeing "<optimized out>" in the
watchlist for C variables in an optimized build).

Given that it is a requirement that natives work in both framed and frameless
modes, a deterministic method is used to alternate between dispatches that are
framed and frameless in the debug build.  This helps ensure that both code paths
stay exercised.  (See SPORADICALLY(n)).

By alternating framed and frameless variants, differences become more apparent,
and one difference was in the error reporting.  When a framed native reported a
type error about an argument, it appeared not to originate from within the call stack
of the function because it "hadn't run yet".  That same error message coming from
the frameless native would appear to have a call stack with the native running.  A
compensation strategy was chosen to select certain "invalid argument" style errors
to detect if they were originating from a frameless call, and omit that stack level if
so...in order to make the errors equivalent.

Though only a beginning to the exploration of what can be accomplished with the
method, it already has shown the COMMENT function running more than 2x faster
than in R3-Alpha.  Similar gains should be applicable to a fairly broad category of
functions...for which a cost-benefit analysis should be done in order to decide if the
additional code maintenance is warranted for the speedup.